third_party/WebKit/LayoutTests/fast/forms/input-type-change-focusout.html

<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<input id="input" type='date' />
<script>
description("This tests that changing an input element's type does not cause focusout event.");

focusoutCalled = false;

input.addEventListener('focusout', function() {
    focusoutCalled = true;
});
input.focus();
input.type = 'month';
shouldBeFalse('focusoutCalled');
input.type = 'week';
shouldBeFalse('focusoutCalled');
input.type = 'time';
shouldBeFalse('focusoutCalled');
input.type = 'datetime-local';
shouldBeFalse('focusoutCalled');
input.type = 'text';
shouldBeFalse('focusoutCalled');

</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
